Clean compilation... one stubbed constructor.
